fix some isseus

+++ staff/src/migrations/20220222064147-create-permissions.js
@@ -15,7 +15,7 @@ module.exports = {
                 type: Sequelize.STRING
             },
             businessId: {
-                type: Sequelize.NUMERIC
+                type: Sequelize.INTEGER
             },
             roleId: {
                 type: Sequelize.INTEGER,
